[gdal] GDAL: add netCDF dependency (#652)
* GDAL: add netCDF dependency

Note that given the way GDAL will use it
(see https://trac.osgeo.org/gdal/changeset/38914 ), we won't fuzz netCDF files
themselves and test libnetcdf itself. Instead, for the purpose of oss-fuzz,
GDAL will use as input a text file (syntax of the ncdump/ncgen utilitis)
and will convert it to a (valid) netCDF file on-the-fly. This way we test how
the GDAL netCDF driver reacts, instead of potential bug in libnetcdf itself.

* GDAL: compile libnetcdf against libhdf5_serial to enable netcdf-4 support