[mupdf] Plug memory leaks in fuzzer itself. (oss-fuzz #5497) (#1778)
If mupdf throws an exception (extremely likely in when fuzzing),
none of the resources such as the rendered pixmap, document or
stream were reclaimed before exiting causing memory leaks. These
leaks were blamed on the software itself, when it actually was
the fuzzer implementation that leaked.

Fix this by always cleaning up all resources.